+++ ChangeLog.md
@@ -6,8 +6,9 @@ This project's release branch is `master`. This log is written from the perspect
 
 * Generalised pathSegmentEncoder, added pathFieldEncoder.
 * Added some Prisms to the encoder library for manipulating DSums (perhaps they should get moved to dependent-sum before release?)
-* Add "ob doc" command, which lists paths to haddock documentation for specified packages
+* Add `ob doc` command, which lists paths to haddock documentation for specified packages
 * Bump reflex-platform so that obelisk now uses GHC 8.6.5 and the nixos-19.03 nixpkgs set
+* Add local packages from the Nix `project { packages }` option to the GHCi config, so that they are auto-reloaded with `ob run`.
 
 ## v0.2.0.0 - 2019-8-17